Adding Base Entity
