+++ src/main/java/com/non/model/User.java
@@ -1,22 +1,20 @@
 package com.non.model;
 
+import com.non.util.Role;
 import jakarta.persistence.*;
 import lombok.Data;
-
-import java.sql.Timestamp;
-import java.util.UUID;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
 
 
 @Entity
 @Data
-public class User {
+@EqualsAndHashCode(callSuper=false)
+@NoArgsConstructor
+public class User extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.UUID)
-    private UUID id;
-    private Timestamp createTime;
-    private Timestamp updateTime;
-    private Integer createdBy;
-    private Integer updatedBy;
+    private String id;
 
     @Column(unique = true, nullable = false)
     private String username;
@@ -32,12 +30,9 @@ public class User {
     private boolean isActive = true;
 
     @Enumerated(EnumType.STRING)
-    private ROLE role;
+    private Role role;
 
-    public User() {
-    }
-
-    public User(String username, String firstName, String lastName, String email, ROLE role) {
+    public User(String username, String firstName, String lastName, String email, Role role) {
         this.username = username;
         this.firstName = firstName;
         this.lastName = lastName;
@@ -50,108 +45,6 @@ public class User {
         this.firstName = firstName;
         this.lastName = lastName;
         this.email = email;
-        this.role = ROLE.MENTEE;
-    }
-
-    public UUID getId() {
-        return id;
-    }
-
-    public void setId(UUID id) {
-        this.id = id;
-    }
-
-    public Timestamp getCreateTime() {
-        return createTime;
-    }
-
-    public void setCreateTime(Timestamp createTime) {
-        this.createTime = createTime;
-    }
-
-    public Timestamp getUpdateTime() {
-        return updateTime;
-    }
-
-    public void setUpdateTime(Timestamp updateTime) {
-        this.updateTime = updateTime;
-    }
-
-    public Integer getCreatedBy() {
-        return createdBy;
-    }
-
-    public void setCreatedBy(Integer createdBy) {
-        this.createdBy = createdBy;
-    }
-
-    public Integer getUpdatedBy() {
-        return updatedBy;
-    }
-
-    public void setUpdatedBy(Integer updatedBy) {
-        this.updatedBy = updatedBy;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public boolean isDeleted() {
-        return isDeleted;
-    }
-
-    public void setDeleted(boolean deleted) {
-        isDeleted = deleted;
-    }
-
-    public boolean isActive() {
-        return isActive;
-    }
-
-    public void setActive(boolean active) {
-        isActive = active;
-    }
-
-    public ROLE getRole() {
-        return role;
-    }
-
-    public void setRole(ROLE role) {
-        this.role = role;
-    }
-
-    public enum ROLE {
-        ADMIN,
-        MENTOR,
-        MENTEE
+        this.role = Role.MENTEE;
     }
 }
