+++ src/main/java/com/non/service/impl/UserServiceImpl.java
@@ -1,0 +1,24 @@
+package com.non.service.impl;
+
+import com.non.model.User;
+import com.non.repository.UserRepository;
+import com.non.service.UserService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class UserServiceImpl implements UserService {
+    private final UserRepository userRepository;
+
+    @Autowired
+    public UserServiceImpl(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
+
+    @Override
+    public List<User> getAllUsers() {
+        return userRepository.findAll();
+    }
+}
