+++ src/main/java/com/non/util/Role.java
@@ -1,0 +1,7 @@
+package com.non.util;
+
+public enum Role {
+    ADMIN,
+    MENTOR,
+    MENTEE
+}
