+++ src/main/resources/application.properties
@@ -4,7 +4,7 @@ spring.datasource.username=root
 spring.datasource.password=root
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQLDialect
 spring.jpa.generate-ddl=true
-spring.jpa.hibernate.ddl-auto=update
+spring.jpa.hibernate.ddl-auto=create-drop
 spring.datasource.tomcat.test-on-borrow=true
 spring.datasource.tomcat.validation-query=SELECT 1
 # JPA (JpaBaseConfiguration, HibernateJpaAutoConfiguration)
@@ -24,10 +24,8 @@ springdoc.swagger-ui.tryItOutEnabled=true
 springdoc.swagger-ui.filter=true
 non.openapi.dev-url=http://localhost:8080
 non.openapi.prod-url=https://non-api.com
-
 # Reduce logging level. Set logging level to info
 logging.level.root=info
-
 # Show JPA/Hibernate logging messages
 logging.level.org.hibernate.SQL=trace
 logging.level.org.hibernate.orm.jdbc.bind=trace
